Expose gaps commands as library
Expose init, download, and upload functions so gaps can be used in other
javascript programs.

Update clone and download functions to take an optional destination
directory.  Update download function to return the project object so the
path to the actual project, based on the project's name, can be
retrieved.

Update upload function to take an optional source directory.  When
uploading, set path in the manifest to the actual path in case the
directory has been moved and the path inside the manifest file is no
longer correct.