Add current remote file list back into manifestor inputs
The upload command was not passing in the list of external files to
manifestor#build, which meant that every local file was being treated
as new. Drive throws an exception when a file with no ID that also
matches a name in the existing file is included in the manifest.